==> Projects/Final Project/3 - User Search.ipynb <==
# Importing Libraries

import requests
import pandas as pd
from tqdm.auto import tqdm
import sqlalchemy as db
from fuzzywuzzy import fuzz

# Creating the connection with database

Transforming the table into a pandas Dataframe

engine = db.create_engine(f'{DATABASE_URL}')
conn = engine.connect()
df_ingredientes = pd.read_sql('ingredient', con=conn)

# Search for recipes

## User input (ingredients)

user_ingredients = []
while True:
    ingredient = str(input('Ingredientes:'))
    if ingredient == 'ok':
        break
    else:
        user_ingredients.append(ingredient)

user_ingredients

## Returning the ingredient's ids

id_ingredient_user = []
for ingredient in user_ingredients:
    lst_id_ingredient = df_ingredientes.loc[df_ingredientes['ingrediente'] == f'{ingredient}', :]['id_ingrediente'].values
    
    if len(lst_id_ingredient) > 0:
        id_ingredient = lst_id_ingredient[0]
        id_ingredient_user.append(id_ingredient)
    else:
        for unique_ingredient in df_ingredientes['ingrediente']:
                ratio = fuzz.token_set_ratio(ingredient, unique_ingredient)
                if ratio > 85:
                    lst_id_ingredient = df_ingredientes.loc[df_ingredientes['ingrediente'] == f'{unique_ingredient}', :]['id_ingrediente'].values
                    id_ingredient = lst_id_ingredient[0]
                    id_ingredient_user.append(id_ingredient)
                    break

id_ingredient_user = tuple(id_ingredient_user)

## Returning the recipes

engine = db.create_engine(f'{DATABASE_URL}')
conn = engine.connect()
query = f'''SELECT 
                        t2.href,
                        t2.rendimento,
                        t2.tempo_preparo,
                        t2.nome,
                        t2.qtde_ingrediente_usuario,
                        t2.qtd_ingrediente,
                        t2.ingredients_ratio,
                        t2.categoria,
                        rim.image_url
                    FROM
                        (SELECT 
                            *
                        FROM
                            (SELECT 
                            r.href,
                                r.rendimento,
                                r.tempo_preparo,
                                r.nome,
                                COUNT(i.id_ingrediente) AS qtde_ingrediente_usuario,
                                r.qtd_ingrediente,
                                (COUNT(i.id_ingrediente) / r.qtd_ingrediente) AS ingredients_ratio,
                                r.categoria
                        FROM
                            ingredient i
                        JOIN recipe_ingredient ri ON i.id_ingrediente = ri.id_ingredient
                        JOIN recipes_df r ON r.id_recipes_df = ri.id_recipe
                        WHERE
                            i.id_ingrediente IN {id_ingredient_user}
                        GROUP BY ri.id_recipe
                        ORDER BY qtde_ingrediente_usuario DESC) AS t
                        WHERE
                            t.ingredients_ratio > 0.6
                            LIMIT 200) AS t2
                            LEFT JOIN
                        recipes.images rim ON t2.href = rim.href
                        WHERE rim.image_url IS NOT NULL
                        ORDER BY t2.ingredients_ratio DESC
                        LIMIT 21'''

results = engine.execute(query)
recipes = [recipe for recipe in results]

recipes[:5]
